martes 04

+++ src/main/java/GranPremio/Interfaces/Avanzable.java
@@ -2,5 +2,5 @@ package GranPremio.Interfaces;
 
 public interface Avanzable {
 	int calcularAvanceTurno();
-	void aplicarAvance(double metros);
+	void aplicarAvance(int metros);
 }
